04-04-2025 night 6:59
from deepak 24 hours fun

+++ app/src/main/java/com/mykaimeal/planner/fragment/mainfragment/viewmodel/planviewmodel/apiresponsebydate/DataPlayByDate.kt
@@ -12,9 +12,9 @@ data class DataPlayByDate(
     val Teatime: MutableList<BreakfastModelPlanByDate>?,
     val fridge:Int?,
     val freezer:Int?,
-    val fat:Double?,
-    val protein:Double?,
-    val kcal:Double?,
-    val carbs:Double?,
+    val fat:Int?,
+    val protein:Int?,
+    val kcal:Int?,
+    val carbs:Int?,
     var show:Int
 )
